Some tidying up
