+++ src/main/java/BonusMilesService.java
@@ -1,9 +1,9 @@
 public class BonusMilesService {
-    public long calculate ( long amount, boolean registered ) {
+    public long calculate(long amount, boolean registered) {
         int percent = registered ? 3 : 1;
         long bonus = amount * percent / 100 / 100;
         long limit = 500;
-        if (bonus >limit) {
+        if (bonus > limit) {
             bonus = limit;
         }
         return bonus;
